Organization and renamed Handler files to Manager
